requirements.txt
opencv-python
easyocr
numpy
simple-lama-inpainting
pillow
